extract split-column, replace all instances in summon
