+++ apps/summon-app/src/layouts/AdvancedSegment.tsx
@@ -1,6 +1,5 @@
-import { FormSegment, WrappedInput } from '@daohaus/ui';
+import { FormSegment, SplitColumn, WrappedInput } from '@daohaus/ui';
 
-import { SplitColumn } from '../layouts/FormLayouts';
 import { INFO_COPY } from '../utils/content';
 import { FORM_KEYS } from '../utils/formKeys';
