+++ apps/summon-app/src/layouts/FormLayouts.tsx
@@ -1,60 +1,6 @@
 import { widthQuery } from '@daohaus/ui';
 import styled from 'styled-components';
 
-const StyledSplitColumn = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  flex-direction: column;
-`;
-type Row = {
-  rowID: string;
-  left: React.ReactNode;
-  right: React.ReactNode;
-};
-type SplitColumnProps = {
-  singleRow?: Row;
-  rows?: Row[];
-};
-
-export const SplitColumn = ({ rows, singleRow }: SplitColumnProps) => {
-  return (
-    <StyledSplitColumn>
-      {singleRow && <Row {...singleRow} />}
-      {rows?.map((row) => {
-        return <Row key={row.rowID} {...row} />;
-      })}
-    </StyledSplitColumn>
-  );
-};
-
-const StyledRow = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  margin-bottom: 3.6rem;
-  .left-section,
-  .right-section {
-    width: 24rem;
-    @media ${widthQuery.sm} {
-      width: 100%;
-      margin-bottom: 1.6rem;
-    }
-  }
-  @media ${widthQuery.sm} {
-    flex-direction: column;
-  }
-`;
-
-const Row = ({ left, right }: Row) => {
-  return (
-    <StyledRow>
-      <div className="left-section">{left}</div>
-      <div className="right-section">{right}</div>
-    </StyledRow>
-  );
-};
-
 export const TextAreaSection = styled.div`
   width: 100%;
   margin-bottom: 3.4rem;
