+++ libs/ui/src/components/layouts/SplitColumn/index.ts
@@ -1,0 +1,1 @@
+export * from './SplitColumn';
