+++ libs/ui/src/components/layouts/index.ts
@@ -1,2 +1,3 @@
 export * from './OuterLayout';
 export * from './FormSegment';
+export * from './SplitColumn';
